First draft of a clojure front end to drake
Add the clj-frontend ns and convert drake.core/configure-logging into a
public function.

A serious downside to the code in this commit is that some of it is copy
and pasted out of the relevant functions in the parser ns, leading to
code duplication.  This will make it harder to alter the parse tree in
the future because changes will be needed in both parser and
clj-frontend.  A cleaner option would be to rewrite parser so that it
delegates to clj-frontend when it actually wants to modify the
parse-tree.  Unfortunately, I don't understand fnparse well enough to
know if this approach would be possible.