Shorten fn names and reverse infiles and outfiles
Now outfiles come before infiles and the method names are shorter, e.g
method-step instead of add-method-step.

Add a function add-methods which can add methods in bulk to the
parse tree.

Rename opts to options in step, method-step, cmd-step and method.

Rename opts to run-options in run-workflow.

Update the example to reflect some of these changes.