Add with-ns macro to support c4 runtime dependency resolution
Without this macro, if your c4 step needs to use a namespace that
isn't already required, you'd need to do an explicit require and
resolve in order to bring in runtime dependencies not already in
the c4 namespace. This macro makes it a bit easier by allowing
you to specify the namespace(s) and function(s) you need in your
body and the macro takes care of the rest.